Solution for $sudo::conf not loading in $sudo::params prior to parameters being set.
In the $sudo:: conf definition it was defaulting to a $sudo::params value in one of it's
parameters if they were not set. Unfortunately puppet does not load in including classes (For
example `sudo` was included, which inherits from $sudo::params; however this is not included
until *after* those parameters are set.

The solution inclues still allowing the user to set this parameter, which it defaults to undef. There
is then a conditional that checks if it is undefined, and then sets it to the $sudo::params value - and if
it is defined, it sets it to what was set.