require tosmi/puppet-ldap, changed uri to http
this fork contains support for AIX. otherwise the rspec tests for the
aix package installation fail.

there's a pull request open for the upstream module torian/puppet-ldap,
as soon as this gets merged we can change the requirement to this
module.

i've also changed the github uri's to http so that cloning the required
repos behind a http proxy works.