refactor rspec tests
if you look closely at the travis-ci output the only test run before
this commit was the check for class[sudo::params]. all other tests were
skipped. i'm not an rspec expert to clearly state why, but with this
commit all tests get executed.

added a test for the sudo::package class for redhat and debian.

also added a fact for the AIX tests (osfamily) as it was missing.