move package_ensure to params.pp
and set it to 'latest' on redhat releases 5.[01234].

redhat < 5.5 comes with sudo 1.6 which does not support the #includedir
directive. so we have to make sure we run the latest sudo version.
this will automatically update sudo to 1.7.2 on older releases.

nevertheless you have to have an updated version of sudo
available (e.g. satellite, rhn or a custom repo).