conf: Fix the scope of the visudo -c check
Currently, the module does a visudo -c - which check everyfile - to
validate that the sudo::conf was correct. Unfortunately, if one needs
to be in purge -> false, and some files there have wrong permissions
because of a packaging related issues, it cancels the sudo::conf even
if the rule is perfectly valid. This patch aims to fix that, using the
-f option of visudo to validate only the file we are adding.